requirements.txt
asgiref==3.10.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
lxml==6.0.2
numpy==2.3.4
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pycparser==2.23
PyMuPDF==1.22.3
pypdfium2==5.0.0
pytesseract==0.3.13
python-docx==1.2.0
sqlparse==0.5.3
typing_extensions==4.15.0
